+++ test/simple/cppnet_client.cpp
@@ -4,8 +4,6 @@
 
 #include "include/cppnet.h"
 #include "common/util/time.h"
-#include "include/cppnet_socket.h"
-
 
 using namespace cppnet;
 
@@ -41,7 +39,7 @@ void ReadFunc(Handle handle, std::shared_ptr<Buffer> data, uint32_t len) {
     }
 
     auto msg = GetMsg();
-    handle->Write(msg.c_str(), msg.length());
+    handle->Write(msg.c_str(), (uint32_t)msg.length());
 }
 
 void ConnectFunc(Handle handle, uint32_t err) {
@@ -51,7 +49,7 @@ void ConnectFunc(Handle handle, uint32_t err) {
         handle->GetAddress(ip, port);
         std::cout << " [ConnectFunc] : ip : " << ip << "port : " << port << std::endl;
         auto msg = GetMsg();
-        handle->Write(msg.c_str(), msg.length());
+        handle->Write(msg.c_str(), (uint32_t)msg.length());
 
     } else {
         std::cout << " [ConnectFunc] some thing error : " << err << std::endl;
